Extract Message Types component
